@@ -1,13 +1,16 @@
 // /docker/web-app/src/app/MainLayout.tsx
 'use client';
 
-import Sidebar from '@/components/Sidebar';
+import Sidebar from '../components/Sidebar';
 import { usePathname } from 'next/navigation';
 import clsx from 'clsx';
 
+export const shouldHideSidebar = (pathname: string) =>
+  /^\/dashboard\/(dam-explorer|explorer|motion|live|witness)/.test(pathname);
+
 export default function MainLayout({ children }: { children: React.ReactNode }) {
   const pathname = usePathname();
-  const hideSidebar = /^\/dashboard\/(camera-monitor|dam-explorer|explorer|motion|live|witness)/.test(pathname);
+  const hideSidebar = shouldHideSidebar(pathname);
 
   return (
     <div className="flex flex-1 overflow-hidden">
