@@ -10,6 +10,7 @@ import VideoSocketProvider from './VideoSocketProvider'
 import AssetProvider       from './AssetProvider'
 import ModalProvider       from './ModalProvider'
 import ActionSheet         from '@/components/modals/ActionSheet'
+import { SidebarProvider } from '../hooks/useSidebar'
 
 const qc = new QueryClient()
 
@@ -30,18 +31,20 @@ const CaptureProvider = dynamic(
 export default function AppProviders({ children }: { children: ReactNode }) {
   return (
     <QueryClientProvider client={qc}>
-      <ThemeProvider>
-        <VideoSocketProvider>
-          <CaptureProvider>
-            <AssetProvider>
-              <ModalProvider>
-                {children}
-                <ActionSheet />
-              </ModalProvider>
-            </AssetProvider>
-          </CaptureProvider>
-        </VideoSocketProvider>
-      </ThemeProvider>
+      <SidebarProvider>
+        <ThemeProvider>
+          <VideoSocketProvider>
+            <CaptureProvider>
+              <AssetProvider>
+                <ModalProvider>
+                  {children}
+                  <ActionSheet />
+                </ModalProvider>
+              </AssetProvider>
+            </CaptureProvider>
+          </VideoSocketProvider>
+        </ThemeProvider>
+      </SidebarProvider>
       <ReactQueryDevtools initialIsOpen={false} />
     </QueryClientProvider>
   )
