@@ -1,31 +1,17 @@
 // /host/services/capture-daemon/scanner/csi_or_usb.go
 package scanner
 
+// Legacy wrappers that forward to the shared scanner helpers. These keep the
+// old function names so existing code continues to compile.
+
 import (
 	"fmt"
-	"os"
-	"path/filepath"
-	"syscall"
-	"unsafe"
+
+	sharedscanner "github.com/Cdaprod/ThatDamToolbox/host/services/shared/scanner"
 )
 
-// these constants come from <linux/videodev2.h>
-const (
-	VIDIOC_QUERYCAP        = 0x80685600
-	V4L2_CAP_VIDEO_CAPTURE = 0x00000001
-)
-
-type v4l2_capability struct {
-	Driver       [16]byte
-	Card         [32]byte
-	BusInfo      [32]byte
-	Version      uint32
-	Capabilities uint32
-	Reserved     [4]uint32
-}
-
-// findFirstCaptureDevice returns the first /dev/videoN that
-// actually has V4L2_CAP_VIDEO_CAPTURE set.
+// findFirstCaptureDevice returns the first /dev/videoN that has capture
+// capabilities by delegating to the shared helpers.
 func findFirstCaptureDevice() (string, error) {
 	for _, node := range sortedVideoNodes() {
 		if IsCaptureNode(node) {
@@ -35,36 +21,8 @@ func findFirstCaptureDevice() (string, error) {
 	return "", fmt.Errorf("no video capture devices found")
 }
 
-// sortedVideoNodes returns a deterministic list of "/dev/video*" sorted by N
-func sortedVideoNodes() []string {
-	matches, _ := filepath.Glob("/dev/video*")
-	// you can add a sort.Slice here if order matters
-	return matches
-}
+// sortedVideoNodes returns a deterministic list of "/dev/video*" nodes.
+func sortedVideoNodes() []string { return sharedscanner.SortedVideoNodes() }
 
-// IsCaptureNode does the VIDIOC_QUERYCAP ioctl and checks the CAP flag
-func IsCaptureNode(dev string) bool {
-	f, err := os.OpenFile(dev, os.O_RDONLY|syscall.O_NONBLOCK, 0)
-	if err != nil {
-		return false
-	}
-	defer f.Close()
-
-	var caps v4l2_capability
-	_, _, errno := syscall.Syscall(syscall.SYS_IOCTL, f.Fd(),
-		VIDIOC_QUERYCAP, uintptr(unsafe.Pointer(&caps)))
-	if errno != 0 {
-		return false
-	}
-	return caps.Capabilities&V4L2_CAP_VIDEO_CAPTURE != 0
-}
-
-func main() {
-	dev, err := findFirstCaptureDevice()
-	if err != nil {
-		fmt.Println("❌", err)
-		os.Exit(1)
-	}
-	fmt.Println("✅ using video device:", dev)
-	// ... now pass `dev` into your FFmpeg command builder
-}
+// IsCaptureNode checks if the given node exposes capture capabilities.
+func IsCaptureNode(dev string) bool { return sharedscanner.IsCaptureNode(dev) }
