@@ -1,45 +1,24 @@
 // /host/services/capture-daemon/scanner/scanner.go
 package scanner
 
-import "time"
+// This package now wraps the shared scanner implementation so existing imports
+// continue to function while discovery logic lives in host/services/shared.
+
+import sharedscanner "github.com/Cdaprod/ThatDamToolbox/host/services/shared/scanner"
 
 // Device represents everything we know about a discovered capture device.
-type Device struct {
-	ID           string                 `json:"id"`             // was UID
-	Kind         string                 `json:"kind,omitempty"` // "v4l2", "usb", "network", etc.
-	Path         string                 `json:"path,omitempty"` // e.g. "/dev/video0" or "rtsp://..."
-	Name         string                 `json:"name"`           // human-readable
-	Capabilities map[string]interface{} `json:"capabilities"`   // whatever the scanner filled in
-	Status       string                 `json:"status,omitempty"`
-	LastSeen     time.Time              `json:"last_seen"`
-}
+// It aliases the shared Device type for backward compatibility.
+type Device = sharedscanner.Device
 
 // Scanner is implemented by each discovery module.
-type Scanner interface {
-	Scan() ([]Device, error)
-}
-
-var scanners []Scanner
+type Scanner = sharedscanner.Scanner
 
 // Register adds a new discovery backend (called in each scanner’s init).
 func Register(s Scanner) {
-	scanners = append(scanners, s)
+	sharedscanner.Register(s)
 }
 
-// ScanAll runs every scanner and merges their results.
+// ScanAll runs every registered scanner and merges their results.
 func ScanAll() ([]Device, error) {
-	var all []Device
-	now := time.Now()
-	for _, s := range scanners {
-		devs, err := s.Scan()
-		if err != nil {
-			// optionally log
-			continue
-		}
-		for _, d := range devs {
-			d.LastSeen = now
-			all = append(all, d)
-		}
-	}
-	return all, nil
+	return sharedscanner.ScanAll()
 }
