@@ -1,39 +1,6 @@
-// /host/services/capture-daemon/scanner/system.go
+// Package scanner previously implemented a local system scanner. The logic now
+// lives in the shared module; this file remains so existing imports continue to
+// compile and to pull in the shared V4L2 scanner via blank import.
 package scanner
 
-import (
-	"fmt"
-	"path/filepath"
-)
-
-// auto-register this backend at init-time
-func init() {
-	Register(systemScanner{})
-}
-
-type systemScanner struct{}
-
-// Scan discovers the first usable /dev/video* node reported by the
-// shared helpers in csi_or_usb.go and wraps it in a Device.
-func (systemScanner) Scan() ([]Device, error) {
-	var out []Device
-	for _, node := range sortedVideoNodes() {
-		if !IsCaptureNode(node) { // <-- new guard
-			continue
-		}
-		out = append(out, Device{
-			ID:   node,
-			Kind: "system",
-			Path: node,
-			Name: filepath.Base(node),
-			Capabilities: map[string]interface{}{
-				"source": "system",
-			},
-			Status: "online",
-		})
-	}
-	if len(out) == 0 {
-		return nil, fmt.Errorf("no usable capture devices")
-	}
-	return out, nil
-}
+import _ "github.com/Cdaprod/ThatDamToolbox/host/services/shared/scanner/v4l2"
